Quick fix

+++ app/src/main/java/com/example/cookme/addItem.java
@@ -111,7 +111,7 @@ public class addItem extends AppCompatActivity {
 
                                 if(key!=-1)
                                 {
-                                    long currAmount = ingredientList.get(key).getAmount();
+                                    long currAmount = (long)ingredientList.get(key).getAmount();
                                     long newAmount = currAmount + LAmount;
                                     //update data syntax:
                                     //mDatabase.child("users").child(userId).child("username").setValue(name);
